Create colonyMotions utils and extract there helpers
